Move OSAgentHijacker to utils
